feat(configuration): Allow configuration of dynamic names for the socket.
Fix #8.
